+++ common/src/main/java/net/mehvahdjukaar/supplementaries/client/renderers/tiles/NoticeBoardBlockTileRenderer.java
@@ -211,14 +211,13 @@ public class NoticeBoardBlockTileRenderer implements BlockEntityRenderer<NoticeB
         if (pattern != null) {
             VertexConsumer builder = pattern.buffer(buffer, RenderType::entityTranslucent);
             int i =  tile.getDyeColor().getTextColor();
-            float scale = 0.8f;//so its more similar to text. idk why its needed
+            float scale = 0.5f;//so its more similar to text. idk why its needed
             int b = (int) (scale* (FastColor.ARGB32.blue(i)));
             int g = (int) (scale*(FastColor.ARGB32.green(i)));
             int r = (int) (scale*(FastColor.ARGB32.red(i)));
             int lu = frontLight & '\uffff';
             int lv = frontLight >> 16 & '\uffff';
-            poseStack.mulPose(RotHlpr.Y180);
-            poseStack.translate(0,0,-0.008f);
+            poseStack.translate(0,0,0.008f);
             VertexUtil.addQuad(builder, poseStack, -0.4375F, -0.4375F,  0.4375F, 0.4375F,
                     0.15625f, 0.0625f, 0.5f + 0.09375f, 1 - 0.0625f, r, g, b, 255, lu, lv);